# :bug: :racehorse: Fix ROM generation on specific stream and improve performance