--- a/.vscode/settings.json
+++ b/.vscode/settings.json
@@ -16,6 +16,7 @@
     "cSpell.words": [
         "Agcb",
         "cref",
+        "Diagnoser",
         "Ekona",
         "ESRB",
         "HMAC",