--- a/src/Directory.Packages.props
+++ b/src/Directory.Packages.props
@@ -1,14 +1,14 @@
 <Project>
     <!-- Centralize dependency management -->
     <ItemGroup>
-        <PackageVersion Include="Yarhl" Version="4.0.0-preview.153" />
+        <PackageVersion Include="Yarhl" Version="4.0.0-preview.157" />
         <PackageVersion Include="Texim" Version="0.1.0-preview.129" />
         <PackageVersion Include="System.Data.HashFunction.CRC" Version="2.0.0" />
         <PackageVersion Include="Portable.BouncyCastle" Version="1.9.0" />
 
         <PackageVersion Include="YamlDotNet" Version="11.2.1" />
-        <PackageVersion Include="FluentAssertions" Version="6.5.1" />
-        <PackageVersion Include="nunit" Version="3.13.2" />
+        <PackageVersion Include="FluentAssertions" Version="6.6.0" />
+        <PackageVersion Include="nunit" Version="3.13.3" />
         <PackageVersion Include="NUnit3TestAdapter" Version="4.2.1" />
         <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.1.0"/>
         <PackageVersion Include="coverlet.collector" Version="3.1.2" />
@@ -17,7 +17,7 @@
         <PackageVersion Include="Microsoft.SourceLink.GitHub" Version="1.1.1" />
 
         <PackageVersion Include="StyleCop.Analyzers" Version="1.1.118" />
-        <PackageVersion Include="SonarAnalyzer.CSharp" Version="8.36.1.44192" />
-        <PackageVersion Include="Roslynator.Analyzers" Version="4.0.2" />
+        <PackageVersion Include="SonarAnalyzer.CSharp" Version="8.37.0.45539" />
+        <PackageVersion Include="Roslynator.Analyzers" Version="4.1.0" />
     </ItemGroup>
 </Project>