--- a/src/Ekona.PerformanceTests/Ekona.PerformanceTests.csproj
+++ b/src/Ekona.PerformanceTests/Ekona.PerformanceTests.csproj
@@ -13,6 +13,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Ekona\Ekona.csproj" />
     <PackageReference Include="BenchmarkDotNet" />
+    <PackageReference Include="YamlDotNet" />
   </ItemGroup>
 
 </Project>