--- a/src/Ekona.PerformanceTests/FilePathInfo.cs
+++ b/src/Ekona.PerformanceTests/FilePathInfo.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) 2022 SceneGate
+// Copyright (c) 2022 SceneGate
 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -17,12 +17,13 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-using BenchmarkDotNet.Running;
-
 namespace SceneGate.Ekona.PerformanceTests;
 
-public static class Program
+public class FilePathInfo
 {
-    public static void Main(string[] args) =>
-        BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
+    public FilePathInfo(string path) => Path = path ?? throw new ArgumentNullException(nameof(path));
+
+    public string Path { get; set; }
+
+    public override string ToString() => System.IO.Path.GetFileName(Path);
 }