--- a/src/Ekona/Containers/Rom/NitroRom2Binary.cs
+++ b/src/Ekona/Containers/Rom/NitroRom2Binary.cs
@@ -75,8 +75,6 @@ public class NitroRom2Binary :
         if (source is null)
             throw new ArgumentNullException(nameof(source));
 
-        root = source.Root;
-
         // Binary order is:
         // - Header
         // - Unknown
@@ -93,9 +91,13 @@ public class NitroRom2Binary :
         // - Unknown, probably padding but due to cartridge protocol 3x 0x8000..0x9000
         // - ARM9i
         // - ARM7i
+        root = source.Root;
+
         var binary = new BinaryFormat(initializedOutputStream ?? new DataStream());
-        binary.Stream.Position = 0;
-        writer = new DataWriter(binary.Stream);
+
+        // Create a wrapper so the length starts from 0. We shouldn't dispose it.
+        var writeStream = new DataStream(binary.Stream);
+        writer = new DataWriter(writeStream);
 
         programInfo = GetChildFormatSafe<ProgramInfo>("system/info");
         sectionInfo = new RomSectionInfo {
@@ -182,6 +184,7 @@ public class NitroRom2Binary :
         Node? logoNode = root.Children["system"]?.Children["copyright_logo"];
         if (logoNode is not null) {
             DataStream binaryLogo = logoNode.Stream ?? throw new FormatException("Invalid format for copyright_logo");
+            binaryLogo.Position = 0;
             header.CopyrightLogo = new DataReader(binaryLogo).ReadBytes((int)binaryLogo.Length);
         } else {
             header.CopyrightLogo = new byte[156];
@@ -263,9 +266,9 @@ public class NitroRom2Binary :
         }
 
         writer.Stream.Position = sectionInfo.BannerOffset;
-        GetChildSafe("system/banner")
-            .TransformWith<Banner2Binary>()
-            .Stream!.WriteTo(writer.Stream);
+        var bannerContainer = GetChildFormatSafe<NodeContainerFormat>("system/banner");
+        using var bannerStream = (BinaryFormat)ConvertFormat.With<Banner2Binary>(bannerContainer);
+        bannerStream.Stream!.WriteTo(writer.Stream);
 
         writer.Stream.Position = sectionInfo.BannerOffset;
         sectionInfo.BannerLength = (uint)Binary2Banner.GetSize(writer.Stream);